Domasna 3/crypto-app/requirements.txt
numpy
pandas
scikit-learn
tensorflow
torch
psycopg2-binary
matplotlib
python-dotenv
flask
flask-cors
requests
transformers
cachetools
